Refactor to make adding more formats easier
Resolves #7
